% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Clust}
\alias{Clust}
\title{Clust}
\usage{
Clust(df, G = NULL, k = 1, what = "uncertainty")
}
\arguments{
\item{df}{data frame}

\item{G}{mixture components}

\item{k}{number of clusters}

\item{what}{type of plot from mclust}
}
\description{
Clust
}
\examples{
\dontrun{Clust(df = df, G = 1, k = 1, what = "uncertainty")}
}
