% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris.R
\docType{data}
\name{iris}
\alias{iris}
\title{iris}
\format{
A data frame with 150 rows and 5 columns:
\describe{
  \item{Sepal.Length}{Sepal Length}
  \item{Sepal.Width}{Sepal Width}
  \item{Petal.Length}{Petal Length}
  \item{Petal.Width}{Petal Width}
  \item{Species}{Species}
  ...
}
}
\source{
<https://www.amazon.com/Applied-Multivariate-Statistical-Analysis-6th/dp/0131877151>
}
\usage{
iris
}
\description{
A data frame consisting of 150 flowers and corresponding properties.
}
\keyword{datasets}
