% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myFA}
\alias{myFA}
\title{myFA}
\usage{
myFA(df, factors = 1, rotation = "none")
}
\arguments{
\item{df}{data frame}

\item{factors}{number of factors}

\item{rotation}{type of rotation}
}
\description{
myFA
}
\examples{
\dontrun{myFA(df = df, factors = 1, rotation = "none")}
}
