% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myPrinComp}
\alias{myPrinComp}
\title{myPrinComp}
\usage{
myPrinComp(df, mat)
}
\arguments{
\item{df}{data frame}

\item{mat}{R or S matrix}
}
\description{
myPrinComp
}
\examples{
\dontrun{myPrinComp(df = df, mat = "S")}
}
