% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T4_3.R
\docType{data}
\name{T4_3}
\alias{T4_3}
\title{T4_3}
\format{
A data frame with 30 rows and 5 columns:
\describe{
  \item{V1}{First measurement}
  \item{V2}{Second measurement}
  \item{V3}{Third measurement}
  \item{V4}{Fourth measurement}
  \item{V5}{Statistical Distance}
  ...
}
}
\source{
<https://www.amazon.com/Applied-Multivariate-Statistical-Analysis-6th/dp/0131877151>
}
\usage{
T4_3
}
\description{
A data frame consisting of 30 trees and corresponding stiffness results.
}
\keyword{datasets}
