% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T8_4.R
\docType{data}
\name{T8_4}
\alias{T8_4}
\title{T8_4}
\format{
A data frame with 103 rows and 5 columns:
\describe{
  \item{V1}{JP Morgan}
  \item{V2}{CitiBank}
  \item{V3}{Wells Fargo}
  \item{V4}{Royal Dutch Shell}
  \item{V5}{Exxon Mobil}
  ...
}
}
\source{
<https://www.amazon.com/Applied-Multivariate-Statistical-Analysis-6th/dp/0131877151>
}
\usage{
T8_4
}
\description{
A data frame consisting of 103 stock prices and corresponding banks.
}
\keyword{datasets}
