% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T8_5.R
\docType{data}
\name{T8_5}
\alias{T8_5}
\title{T8_5}
\format{
A data frame with 61 rows and 5 columns:
\describe{
  \item{V1}{Total Population}
  \item{V2}{Profession}
  \item{V3}{Employment}
  \item{V4}{Government Employment}
  \item{V5}{Medium Home Value}
  ...
}
}
\source{
<https://www.amazon.com/Applied-Multivariate-Statistical-Analysis-6th/dp/0131877151>
}
\usage{
T8_5
}
\description{
A data frame consisting of 61 measurements and corresponding banks.
}
\keyword{datasets}
